Request and receive full object report packets

+++ code/arduino/pa_driver/pixart.hpp
@@ -29,7 +29,7 @@ struct PA_object
   }
 };
 
-double PA_get_frame_period_microseconds();
+uint32_t PA_get_frame_period_microseconds();
 void PA_write(uint8_t bank, uint8_t reg, uint8_t data);
 uint8_t PA_read(uint8_t bank, uint8_t reg);
 void PA_read_report(uint8_t buffer[], int format);
